refactor : Ladder의 사다리 게임 진행 책임을 LadderGame으로 분리
