+++ src/main/java/controller/Controller.java
@@ -28,11 +28,10 @@ public class Controller {
         Width width = new Width(users.getCount() - 1);
         List<String> inputResult = settingResult(users.getCount());
         Height height = settingHeight();
-        Ladder ladder = new Ladder(height, width, new RandomLadderGenerator());
+        LadderGame ladderGame = new LadderGame(height, width, new RandomLadderGenerator());
+        showLadder(users, ladderGame.getLadder(), inputResult);
 
-        showLadder(users, ladder, inputResult);
-
-        Result result = playLadderGame(users, ladder, inputResult);
+        final Result result = ladderGame.play(users, inputResult);
         showResult(users, result);
     }
 
@@ -78,10 +77,10 @@ public class Controller {
         outputView.printResult(inputResult);
     }
 
-    private Result playLadderGame(final Users users, final Ladder ladder, List<String> inputResult) {
-        Map<String, Integer> ladderGameResult = ladder.play(users);
-        return new Result(inputResult, ladderGameResult);
-    }
+//    private Result playLadderGame(final Users users, final Ladder ladder, List<String> inputResult) {
+//        Map<String, Integer> ladderGameResult = ladder.play(users);
+//        return new Result(inputResult, ladderGameResult);
+//    }
 
     private void showResult(final Users users, final Result result) {
         try {
